+++ quill/include/quill/detail/BoundedSPSCQueue.h
@@ -154,6 +154,8 @@ public:
   };
 
 public:
+  using handle_t = Handle;
+
   /**
    * Circular Buffer class Constructor
    * @throws on system error
@@ -240,6 +242,7 @@ private:
   /** Local state - modified by either the producer or consumer, but never both **/
   alignas(CACHELINE_SIZE) uint64_t _local_cached_head{0}; /**< cached head index */
   alignas(CACHELINE_SIZE) uint64_t _local_cached_tail{0}; /**< cached tail index */
+  char _pad0[detail::CACHELINE_SIZE - sizeof(uint64_t)] = "\0";
 };
 
 /***/