+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Improve compiler error message when trying to log a non copy constructible user defined type
 * Fix buffer reallocation bug on TimestampFormatter. In previous versions any timestamp format set to 'set_pattern' expanding to a string longer than 32 bytes would cause a crash. ([#24](https://github.com/odygrd/quill/issues/24))
 * Added a latency benchmark, easily extendable for any logger
+* Added QUILL_USE_BOUNDED_QUEUE. In this mode no new queues get allocated but instead log messages get lost. Number of lost messages is reported to stderr.
 
 ## v1.2.3
 * CMake changes to support package installation in conan.